// src/api/generated/model/createGrantRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { CreateGrantRequestScopeType } from './createGrantRequestScopeType';

export interface CreateGrantRequest {
  userId: string;
  scopeType: CreateGrantRequestScopeType;
  namespace?: string;
  datasourceId?: string;
  readOnly?: boolean;
}